+++ Plugin.cs
@@ -19,14 +19,18 @@ using UnityEngine.Rendering;
 
 namespace HeavyItemSCPs
 {
-    [BepInPlugin(MyPluginInfo.PLUGIN_GUID, MyPluginInfo.PLUGIN_NAME, MyPluginInfo.PLUGIN_VERSION)]
+    [BepInPlugin(PLUGIN_GUID, PLUGIN_NAME, PLUGIN_VERSION)]
     [BepInDependency(LethalLib.Plugin.ModGUID)]
     [BepInDependency("SCP500", BepInDependency.DependencyFlags.SoftDependency)]
     public class Plugin : BaseUnityPlugin
     {
+        const string PLUGIN_GUID = "ProjectSCP.HeavyItemSCPs";
+        const string PLUGIN_NAME = "HeavyItemSCPs";
+        const string PLUGIN_VERSION = "1.0.2";
+
         public static Plugin PluginInstance;
         public static ManualLogSource LoggerInstance;
-        private readonly Harmony harmony = new Harmony(MyPluginInfo.PLUGIN_GUID);
+        private readonly Harmony harmony = new Harmony(PLUGIN_GUID);
         public static PlayerControllerB localPlayer { get { return StartOfRound.Instance.localPlayerController; } }
 
         public static AssetBundle? ModAssets;
@@ -37,6 +41,7 @@ namespace HeavyItemSCPs
         public static ConfigEntry<int> config427MaxValue;
         public static ConfigEntry<string> config427LevelRarities;
         public static ConfigEntry<string> config427CustomLevelRarities;
+        public static ConfigEntry<int> config427SCPDungeonRarity;
 
         public static ConfigEntry<bool> configEnableOpenNecklace;
         public static ConfigEntry<float> configTimeToTransform;
@@ -50,12 +55,14 @@ namespace HeavyItemSCPs
         public static ConfigEntry<int> configMaxSpawns;
         public static ConfigEntry<bool> configIncreaseTimeWhenInInventory;
         public static ConfigEntry<bool> config427_500Compatibility;
+        public static ConfigEntry<bool> config427ResetTransformTimeAtOrbit;
 
         // SCP-427-1 Enemy Configs
         public static ConfigEntry<bool> configEnableSCP4271Transformation;
         public static ConfigEntry<bool> configEnableSCP4271Spawning;
         public static ConfigEntry<string> config4271LevelRarities;
         public static ConfigEntry<string> config4271CustomLevelRarities;
+        public static ConfigEntry<int> config4271SCPDungeonRarity;
 
         // SCP-178 Configs
         public static ConfigEntry<bool> configEnableSCP178;
@@ -96,26 +103,29 @@ namespace HeavyItemSCPs
             config427MinValue = Config.Bind("SCP-427", "Minimum value", 100, "The minimum value of SCP-427.");
             config427MaxValue = Config.Bind("SCP-427", "Maximum value", 150, "The maximum value of SCP-427.");
             config427LevelRarities = Config.Bind("SCP-427 Rarities", "Level Rarities", "ExperimentationLevel:5, AssuranceLevel:7, VowLevel:10, OffenseLevel:15, AdamanceLevel:25, MarchLevel:15, RendLevel:30, DineLevel:35, TitanLevel:45, ArtificeLevel:25, EmbrionLevel:50, All:10, Modded:15", "Rarities for each level. See default for formatting.");
-            config427CustomLevelRarities = Config.Bind("SCP-427 Rarities", "Custom Level Rarities", "", "Rarities for modded levels. Same formatting as level rarities.");
+            config427CustomLevelRarities = Config.Bind("SCP-427 Rarities", "Custom Level Rarities", "Secret LabsLevel:100", "Rarities for modded levels. Same formatting as level rarities.");
+            config427SCPDungeonRarity = Config.Bind("SCP-427 Rarities", "SCP Dungeon Rarity", 100, "The rarity of SCP-427 in the SCP Dungeon. Set to -1 to use level rarities.");
 
             configEnableOpenNecklace = Config.Bind("SCP-427", "Enable open necklace", true, "Whether or not players can open the necklace.");
             configTimeToTransform = Config.Bind("SCP-427", "Time to transform", 120f, "How long a player can hold the necklace before they transform into SCP-427-1. Set to -1 to disable transforming.");
             configTransformOpenMultiplier = Config.Bind("SCP-427", "Transform open multiplier", 3f, "The multiplier applied to the time it takes to transform when the necklace is opened. configTimeToTransform will be multiplied by this value.");
-            configHealthPerSecondHolding = Config.Bind("SCP-427", "Health per second holding", 10, "The health gained per second while holding SCP-427.");
-            configHealthPerSecondOpen = Config.Bind("SCP-427", "Health per second open", 15, "The health gained per second while opening SCP-427.");
-            configHoarderBugTransformTime = Config.Bind("SCP-427", "Hoarder bug transform time", 20f, "The time it takes for the hoarder bug to transform into SCP-427-1. Set to -1 to disable transforming.");
+            configHealthPerSecondHolding = Config.Bind("SCP-427", "Health per second holding", 5, "The health gained per second while holding SCP-427.");
+            configHealthPerSecondOpen = Config.Bind("SCP-427", "Health per second open", 10, "The health gained per second while opening SCP-427.");
+            configHoarderBugTransformTime = Config.Bind("SCP-427", "Hoarder bug transform time", 15f, "The time it takes for the hoarder bug to transform into SCP-427-1. Set to -1 to disable transforming.");
             configBaboonHawkTransformTime = Config.Bind("SCP-427", "Baboon hawk transform time", 10f, "The time it takes for the baboon hawk to transform into SCP-427-1. Set to -1 to disable transforming.");
             configTimeToSpawnSCP4271 = Config.Bind("SCP-427", "Time to spawn SCP-427-1", 300f, "The time it takes for SCP-427 to spawn SCP-427-1 when on the ground. Set to -1 to disable spawning from necklace.");
             configContinueSpawningAfterPickup = Config.Bind("SCP-427", "Continue spawning after picked up", false, "Whether or not SCP-427 should continue spawning SCP-427-1 instances after being picked up by a player or lootbug. Setting this to true will make SCP-427 keep spawning SCP-427-1 instances if its on the ground, regardless of whether its been picked up or not."); // TODO: Set to true for testing, change back later
             configMaxSpawns = Config.Bind("SCP-427", "Max spawns", 1, "The maximum number of SCP-427-1 instances that can be spawned when SCP-427 is on the ground.");
             configIncreaseTimeWhenInInventory = Config.Bind("SCP-427", "Increase time when in inventory", true, "Whether or not timeHoldingSCP427 should increase when SCP-427 is on hotbar but not being held.");
             config427_500Compatibility = Config.Bind("SCP-427", "SCP-500 compatibility", true, "Whether or not SCP-427 should be compatible with the SCP-500 mod. This will only work if you have the SCP-500 mod installed. If enabled, it will temporarily halt the transformation timer when holding or using SCP-427 when you take SCP-500.");
+            config427ResetTransformTimeAtOrbit = Config.Bind("SCP-427", "Reset transform time at orbit", false, "Whether or not SCP-427 should reset its transformation timer when orbiting.");
 
             // SCP-427-1
             configEnableSCP4271Transformation = Config.Bind("SCP-427-1", "Enable SCP-427-1 transformation", true, "Whether or not players or monsters can turn into SCP-427-1.");
             configEnableSCP4271Spawning = Config.Bind("SCP-427-1", "Enable SCP-427-1 spawning", false, "Whether or not SCP-427-1 can spawn naturally from vents.");
             config4271LevelRarities = Config.Bind("SCP-427-1 Rarities", "Level Rarities", "ExperimentationLevel:40, AssuranceLevel:40, VowLevel:40, OffenseLevel:60, AdamanceLevel:60, MarchLevel:60, RendLevel:25, DineLevel:25, TitanLevel:65, ArtificeLevel:70, EmbrionLevel:75, All:30, Modded:30", "Rarities for each level. See default for formatting.");
             config4271CustomLevelRarities = Config.Bind("SCP-427-1 Rarities", "Custom Level Rarities", "", "Rarities for modded levels. Same formatting as level rarities.");
+            config4271SCPDungeonRarity = Config.Bind("SCP-427-1 Rarities", "SCP Dungeon Rarity", 100, "The rarity of SCP-427-1 in the SCP Dungeon. Set to -1 to use level rarities.");
 
             // SCP-178
             /*configEnableSCP178 = Config.Bind("SCP-178", "Enable SCP-178", true, "Whether or not SCP-178 can spawn as scrap.");
@@ -151,23 +161,23 @@ namespace HeavyItemSCPs
             // SCP-427
             if (configEnableSCP427.Value)
             {
-                SCPItems.Load("SCP427", ObjectClass.Safe, config427LevelRarities.Value, config427CustomLevelRarities.Value, config427MinValue.Value, config427MaxValue.Value, 1);
-                SCPItems.LoadEnemy("SCP4271", "SCP427", configEnableSCP4271Spawning.Value);
+                SCPItems.Load("SCP427", ObjectClass.Safe, config427LevelRarities.Value, config427CustomLevelRarities.Value, config427SCPDungeonRarity.Value, config427MinValue.Value, config427MaxValue.Value, 1);
+                SCPItems.LoadEnemy("SCP4271", "SCP427", config4271LevelRarities.Value, config4271CustomLevelRarities.Value, config4271SCPDungeonRarity.Value, configEnableSCP4271Spawning.Value);
             }
 
             // SCP-178
             /*if (configEnableSCP178.Value)
             {
-                SCPItems.Load("SCP178", ObjectClass.Safe, config178LevelRarities.Value, config178CustomLevelRarities.Value, config178MinValue.Value, config178MaxValue.Value, 1);
+                SCPItems.Load("SCP178", ObjectClass.Safe, config178LevelRarities.Value, config178CustomLevelRarities.Value, default, config178MinValue.Value, config178MaxValue.Value, 1);
                 SCPItems.LoadEnemy("SCP1781", "SCP178", false);
                 SCP1783DVision.Load();
             }*/
 
             // Finished
-            Logger.LogInfo($"{MyPluginInfo.PLUGIN_GUID} v{MyPluginInfo.PLUGIN_VERSION} has loaded!");
+            Logger.LogInfo($"{PLUGIN_GUID} v{PLUGIN_VERSION} has loaded!");
         }
 
-        public Dictionary<Levels.LevelTypes, int> GetLevelRarities(string levelsString)
+        public static Dictionary<Levels.LevelTypes, int> GetLevelRarities(string levelsString)
         {
             try
             {
@@ -198,12 +208,12 @@ namespace HeavyItemSCPs
             }
             catch (Exception e)
             {
-                Logger.LogError($"Error: {e}");
+                LoggerInstance.LogError($"Error: {e}");
                 return null;
             }
         }
 
-        public Dictionary<string, int> GetCustomLevelRarities(string levelsString)
+        public static Dictionary<string, int> GetCustomLevelRarities(string levelsString)
         {
             try
             {
@@ -234,7 +244,7 @@ namespace HeavyItemSCPs
             }
             catch (Exception e)
             {
-                Logger.LogError($"Error: {e}");
+                LoggerInstance.LogError($"Error: {e}");
                 return null;
             }
         }