+++ SCPItems.cs
@@ -13,10 +13,10 @@ namespace HeavyItemSCPs
     public static class SCPItems
     {
         public static List<SCPItem> SCPItemsList = new List<SCPItem>();
-        public static List<EnemyType> SCPEnemiesList = new List<EnemyType>();
+        public static List<SCPEnemy> SCPEnemiesList = new List<SCPEnemy>();
 
 
-        public static void Load(string pathName, ObjectClass objClass, string? levelRarities = null, string? customLevelRarities = null, int minValue = 0, int maxValue = 0, int? maxSpawns = null, bool tabletop = false, bool general = false, bool small = false)
+        public static void Load(string pathName, ObjectClass objClass, string? levelRarities = null, string? customLevelRarities = null, int scpDungeonRarity = 0, int minValue = 0, int maxValue = 0, int? maxSpawns = null, bool tabletop = false, bool general = false, bool small = false)
         {
             Item SCP = ModAssets.LoadAsset<Item>($"Assets/ModAssets/{pathName}/{pathName}Item.asset");
             if (SCP == null) { LoggerInstance.LogError($"Error: Couldnt get {pathName} from assets"); return; }
@@ -35,14 +35,14 @@ namespace HeavyItemSCPs
             scp.TabletopItem = tabletop;
             scp.GeneralItem = general;
             scp.SmallItem = small;
-            SCPItemsList.Add(scp);
+            scp.SCPDungeonRarity = scpDungeonRarity;
 
             Dictionary<Levels.LevelTypes, int> levelRaritiesDict = new Dictionary<Levels.LevelTypes, int>();
             Dictionary<string, int> customLevelRaritiesDict = new Dictionary<string, int>();
 
             if (levelRarities != null)
             {
-                string[] levels = levelRarities.Split(',');
+                /*string[] levels = levelRarities.Split(',');
 
                 foreach (string level in levels)
                 {
@@ -59,11 +59,12 @@ namespace HeavyItemSCPs
                     {
                         LoggerInstance.LogError($"Error: Invalid level rarity: {levelType}:{levelRarity}");
                     }
-                }
+                }*/
+                levelRaritiesDict = GetLevelRarities(levelRarities);
             }
             if (customLevelRarities != null)
             {
-                string[] levels = customLevelRarities.Split(',');
+                /*string[] levels = customLevelRarities.Split(',');
 
                 foreach (string level in levels)
                 {
@@ -80,9 +81,11 @@ namespace HeavyItemSCPs
                     {
                         LoggerInstance.LogError($"Error: Invalid level rarity: {levelType}:{levelRarity}");
                     }
-                }
+                }*/
+                customLevelRaritiesDict = GetCustomLevelRarities(customLevelRarities);
             }
 
+            SCPItemsList.Add(scp);
             LethalLib.Modules.NetworkPrefabs.RegisterNetworkPrefab(SCP.spawnPrefab);
             LethalLib.Modules.Utilities.FixMixerGroups(SCP.spawnPrefab);
             LethalLib.Modules.Items.RegisterScrap(SCP, levelRaritiesDict, customLevelRaritiesDict);
@@ -104,7 +107,7 @@ namespace HeavyItemSCPs
             LethalLib.Modules.Items.RegisterShopItem(SCP, price);
         }
 
-        public static void LoadEnemy(string enemyPath, string itemPath, bool enableSpawning = true)
+        public static void LoadEnemy(string enemyPath, string itemPath, string? levelRarities = null, string? customLevelRarities = null, int scpDungeonRarity = 0, bool enableSpawning = true)
         {
             EnemyType enemy = ModAssets.LoadAsset<EnemyType>($"Assets/ModAssets/{itemPath}/{enemyPath}Enemy.asset");
             if (enemy == null) { LoggerInstance.LogError($"Error: Couldnt get {enemyPath} from assets"); return; }
@@ -115,16 +118,41 @@ namespace HeavyItemSCPs
             if (terminalNode == null) { LoggerInstance.LogError($"Error: Couldnt get {enemyPath}TN from assets"); return; }
             if (terminalKeyword == null) { LoggerInstance.LogError($"Error: Couldnt get {enemyPath}TK from assets"); return; }
 
-            SCPEnemiesList.Add(enemy);
+            SCPEnemy scpEnemy = new SCPEnemy();
+            scpEnemy.enemyType = enemy;
+            scpEnemy.SCPDungeonRarity = scpDungeonRarity;
+
+            Dictionary<Levels.LevelTypes, int> levelRaritiesDict = new Dictionary<Levels.LevelTypes, int>();
+            Dictionary<string, int> customLevelRaritiesDict = new Dictionary<string, int>();
+
+            if (levelRarities != null)
+            {
+                levelRaritiesDict = GetLevelRarities(levelRarities);
+            }
+            if (customLevelRarities != null)
+            {
+                customLevelRaritiesDict = GetCustomLevelRarities(customLevelRarities);
+            }
+
+            SCPEnemiesList.Add(scpEnemy);
             LoggerInstance.LogDebug($"Registering {enemyPath} enemy network prefab...");
             LethalLib.Modules.NetworkPrefabs.RegisterNetworkPrefab(enemy.enemyPrefab);
             LoggerInstance.LogDebug($"Registering {enemyPath} enemy...");
-            LethalLib.Modules.Enemies.RegisterEnemy(enemy, 0, Levels.LevelTypes.All, terminalNode, terminalKeyword);
+            LethalLib.Modules.Enemies.RegisterEnemy(enemy, levelRaritiesDict, customLevelRaritiesDict, terminalNode, terminalKeyword);
         }
+    }
+
+    public class SCPEnemy
+    {
+        public string enemyName { get { return enemyType.enemyName; } }
+        public EnemyType enemyType;
+        public int SCPDungeonRarity;
+        public int currentLevelRarity;
     }
 
     public class SCPItem
     {
+        public string itemName { get { return item.itemName; } }
         public Item item;
         public int? MaxSpawns;
         public ObjectClass ObjectClass;
@@ -134,6 +162,9 @@ namespace HeavyItemSCPs
         public bool SmallItem = false;
 
         public int itemsSpawnedInRound = 0;
+
+        public int SCPDungeonRarity;
+        public int currentLevelRarity;
     }
     public enum ObjectClass
     {