+++ ThunderStore/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.2
+- Added a new configs
+- Added Secret Labs level rarity and SCP Dungeon rarity
+- Heal amount for 427 changed from 10 to 5 holding and 15 to 10 when open
+- Fixed some bugs for 427
+- SCP-427-1 now properly spawns when turning spawning on
+
 ## 1.0.1
 - SCP-427-1 now only takes damage from mines if chasing a player
 - Fixed issue with players not transforming